requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1

langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-text-splitters==0.2.4

pypdf==4.3.1
faiss-cpu==1.8.0.post1
tiktoken==0.7.0
